Update vss-extension.json
adding "supportsTasks" property key to make the Trivy task visible in build result summary only when the task is actually being used. We are using Trivy task in few of our pipelines and the tab gives error [Timeline record(s) missing: cannot load results. Is Trivy configured to run on this build?] when trivy is task is not part of the pipeline.