fix: Fix the switching of tabs when there are no results
When the reports are switched between, if the originating report doesn't
have any results the state seems to become stagnent and changing the
drop down doesn't change the tables.

This forces that by using an additional hasReportResults to trigger a
redraw of the VulnerabilitiesTable.

Also switched the tables to take a report rather than a results slice